requirements.txt
fastapi
uvicorn 
python-multipart
google-genai
pymongo
python-docx 
PyPDF2 
python-pptx
python-dotenv
